+++ src/video/pal_video_win32.c
@@ -1213,6 +1213,7 @@ PalResult PAL_CALL palInitVideo(
     s_Video.features2 |= PAL_VIDEO_FEATURE_MONITOR_GET_PRIMARY;
     s_Video.features2 |= PAL_VIDEO_FEATURE_FOREIGN_WINDOWS;
     s_Video.features2 |= PAL_VIDEO_FEATURE_MONITOR_VALIDATE_MODE;
+    s_Video.features2 |= PAL_VIDEO_FEATURE_WINDOW_SET_CURSOR;
 
     s_Video.initialized = true;
     s_Video.allocator = allocator;