+++ tests/tests.lua
@@ -41,7 +41,8 @@ project "tests"
             "input_window_test.c",
             "system_cursor_test.c",
             "attach_window_test.c",
-            "char_event_test.c"
+            "char_event_test.c",
+            "native_integration_test.c"
         }
     end